Wrap observable collection operations in ObservableCollectionManager

+++ src/ValleyAuthenticator/ValleyAuthenticator.Storage/Impl/DirectoryContext.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using ValleyAuthenticator.Storage.Abstract;
 using ValleyAuthenticator.Storage.Abstract.Models;
@@ -51,29 +50,19 @@ namespace ValleyAuthenticator.Storage.Impl
         // Private fields
         private readonly InternalStorageManager _storage;
         private readonly ContextManager _contextManager;
-        private Guid _directoryId;
-        private ObservableCollection<NodeInfo> _collection;
+        private ObservableCollectionManager _collectionManager;
+        private Guid _directoryId;        
         private string _name;
-
-        // Private constants
-        private const string IMAGE_OTP_ENTRY = "key.png";
-        private const string IMAGE_DIRECTORY = "folder.png";
-        
+   
         public DirectoryContext(InternalStorageManager storage, ContextManager contextManager, Guid directoryId)
         {
-            // Input validation
-            if (storage == null)
-                throw new ArgumentNullException(nameof(storage));
-            if (contextManager == null)
-                throw new ArgumentNullException(nameof(contextManager));
             if (directoryId == Guid.Empty)
                 throw new ArgumentException(nameof(directoryId));
 
             // Set private fields
-            _storage = storage;
-            _contextManager = contextManager;
+            _storage = storage ?? throw new ArgumentNullException(nameof(storage));
+            _contextManager = contextManager ?? throw new ArgumentNullException(nameof(contextManager));
             _directoryId = directoryId;
-            _collection = null;
 
             // Query additional information about the directory from the storage
             InternalDirectoryData data = _storage.GetDirectory(directoryId);
@@ -89,100 +78,51 @@ namespace ValleyAuthenticator.Storage.Impl
 
         public ObservableCollection<NodeInfo> ListAndSubscribe()
         {
-            if (_collection == null)
-            {
-                List<NodeInfo> initial = new List<NodeInfo>();
+            if (_collectionManager == null)
+                _collectionManager = ObservableCollectionManager.CreateForDirectory(_storage, _contextManager, _directoryId);
 
-                InternalDirectoryData target = _storage.GetDirectory(_directoryId);
-
-                foreach (InternalDirectoryData directory in target.Directories)
-                {
-                    IDirectoryContext context = _contextManager.GetDirectoryContext(directory.Id);
-
-                    string detail = DisplayUtilities.FormatDirectoryLabel(directory);
-
-                    initial.Add(new NodeInfo(context, directory.Id, directory.Name, detail, IMAGE_DIRECTORY));
-                }
-
-                foreach (InternalOtpEntryData entry in target.OtpEntries)
-                {
-                    IOtpEntryContext context = _contextManager.GetOtpEntryContext(entry.Id);
-
-                    initial.Add(new NodeInfo(context, entry.Id, entry.Data.Issuer, entry.Data.Label, IMAGE_OTP_ENTRY));
-                }
-
-                _collection = new ObservableCollection<NodeInfo>(initial);
-            }
-
-            return _collection;
+            return _collectionManager.Collection;
         }
 
         public Guid AddEntry(OtpData data)
         {
             Guid id = _storage.AddEntry(_directoryId, data);
-
-            IOtpEntryContext context = _contextManager.GetOtpEntryContext(id);
-
-            NodeInfo node = new NodeInfo(context, id, data.Issuer, data.Label, IMAGE_OTP_ENTRY);
-
-            _collection.Add(node);
-
+            _collectionManager?.AddEntry(id);
             return id;
         }
    
         public Guid AddDirectory(string name)
         {
             Guid id = _storage.AddDirectory(_directoryId, name);
-
-            IDirectoryContext context = _contextManager.GetDirectoryContext(id);
-
-            string detail = DisplayUtilities.FormatDirectoryLabel();
-
-            NodeInfo node = new NodeInfo(context, id, name, detail, IMAGE_DIRECTORY);
-
-            // Insert the new directory after the last directory in the collection (before any entries).
-            // The directories are always at the top of the list.
-            int last = 0;
-            while (last < _collection.Count && _collection[last].Context is IDirectoryContext)
-                last++;
-            _collection.Insert(last, node);            
-
+            _collectionManager?.AddDirectory(id);       
             return id;
         }        
 
         public bool Delete()
-        {      
-            Guid? parent = _storage.GetDirectory(_directoryId).Parent;
-            if (!parent.HasValue)
-                throw new Exception("It's not possible to delete the root directory");
-
-            bool deleted = _storage.DeleteDirectory(_directoryId);
-
-            if (deleted)
+        {
+            if (_storage.DirectoryExists(_directoryId))
             {
-                if (_collection != null)
+                Guid? parent = _storage.GetDirectory(_directoryId).Parent;
+                if (!parent.HasValue)
+                    throw new Exception("It's not possible to delete the root directory");
+
+                _storage.DeleteDirectory(_directoryId);
+
+                if (_collectionManager != null)
                 {
-                    _collection.Clear();
-                    _collection = null;
+                    _collectionManager.Clear();
+                    _collectionManager = null;
                 }
 
                 _contextManager.GetDirectoryContext(parent.Value).OnItemDeleted(_directoryId);
-            }           
 
-            return deleted;
+                return true;
+            }
+            return false;
         }
 
         public void OnItemDeleted(Guid id)
-        {
-            for (int i = 0; i < _collection.Count; i++)
-            {
-                if (_collection[i].Id == id)
-                {
-                    _collection.RemoveAt(i);
-                    break;
-                }
-            }
-        }
+            => _collectionManager?.RemoveById(id);
 
         public string GetDetailLabel()
         {
@@ -191,53 +131,7 @@ namespace ValleyAuthenticator.Storage.Impl
         }
 
         public void Validate()
-        {
-            if (_collection != null)
-            {
-                for (int i = 0; i < _collection.Count; i++)
-                {
-                    NodeInfo node = _collection[i];
-                    if (!node.Context.Exists)
-                    {
-                        _collection.RemoveAt(i--);
-                    }
-                    else if (node.Context is IDirectoryContext directoryContext)
-                    {
-                        string detail = directoryContext.GetDetailLabel();
-                        if (node.Detail != detail)
-                            node.UpdateDetail(detail);
-                    }
-                }
-            }
-        }
-
-        private void AddMissingItems()
-        {
-            HashSet<Guid> inCollection = new HashSet<Guid>();
-            foreach (NodeInfo node in _collection)
-                inCollection.Add(node.Id);
-
-            InternalDirectoryData target = _storage.GetDirectory(_directoryId);
-
-            foreach (InternalDirectoryData directory in target.Directories)
-            {
-                if (!inCollection.Contains(directory.Id))
-                {
-                    IDirectoryContext context = _contextManager.GetDirectoryContext(directory.Id);
-                    string detail = DisplayUtilities.FormatDirectoryLabel(directory);
-                    _collection.Add(new NodeInfo(context, directory.Id, directory.Name, detail, IMAGE_DIRECTORY));
-                }
-            }
-
-            foreach (InternalOtpEntryData entry in target.OtpEntries)
-            {
-                if (!inCollection.Contains(entry.Id))
-                { 
-                    IOtpEntryContext context = _contextManager.GetOtpEntryContext(entry.Id);
-                    _collection.Add(new NodeInfo(context, entry.Id, entry.Data.Issuer, entry.Data.Label, IMAGE_OTP_ENTRY));
-                }
-            }
-        }
+            => _collectionManager?.Validate();
 
         public string Export(ExportFormat format)
             => ExportHelper.ExportDirectory(_storage.GetDirectory(_directoryId), format);
@@ -246,7 +140,7 @@ namespace ValleyAuthenticator.Storage.Impl
         {            
             if (ExportHelper.TryImport(_storage, _directoryId, format, data, multiple))
             {
-                AddMissingItems();
+                _collectionManager?.AddMissingNodesForDirectory(_directoryId);
                 return true;
             }
             return false;
